extract new hypothesis form
